pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "streettransformer"
version = "0.1.0"
description = "Vector embeddings for street imagery analysis with DuckDB and FAISS"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Jon Atkins"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "dash>=3.3.0",
    "duckdb>=1.4.2",
    "faiss-cpu>=1.13.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pillow>=10.4.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
faiss = [
    "faiss-cpu>=1.13.0",
]
cli = [
    "open-clip-torch>=3.2.0",
    "torch>=2.0.0",
    "Pillow>=9.0.0,<11.0.0",
    "tqdm>=4.65.0",
    "click>=8.1.0",
]
dashboard = [
    "plotly>=5.17.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
    "mypy>=1.4.0",
]
all = [
    "streettransformer[faiss,cli,dashboard]",
]

[project.scripts]
st-query = "streettransformer.cli.query:main"
st-generate = "streettransformer.cli.generate:main"

[project.urls]
Homepage = "https://github.com/jatkins23/st_dashboard"
Repository = "https://github.com/jatkins23/st_dashboard"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
